Create solution.cpp
You are given a tree of N nodes rooted at 1. Each node of the tree has a colour associated with it. Now you are given Q queries. In each Query, you are given a node number X and for each query you have to mark the node X as special and all the other nodes in its subtree with the same colour also as special. If a node is marked as special in a query then for all the other subsequent queries, it remains marked as special.
For each query, you need to print the total number of special nodes in the tree after you perform the marking operation in the query.